Refactor: lockout by emailLockout
