+++ packages/worker/src/api/routes/global/auth.ts
@@ -2,7 +2,7 @@ import * as authController from "../../controllers/global/auth"
 import { auth } from "@budibase/backend-core"
 import Joi from "joi"
 import { loggedInRoutes } from "../endpointGroups"
-import { lockout } from "../../../middleware"
+import { emailLockout } from "../../../middleware"
 
 function buildAuthValidation() {
   // prettier-ignore
@@ -32,7 +32,7 @@ loggedInRoutes
   .post(
     "/api/global/auth/:tenantId/login",
     buildAuthValidation(),
-    lockout,
+    emailLockout,
     authController.login
   )
   .post("/api/global/auth/logout", authController.logout)
